requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.16.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.2
